backend/requirements.txt
flask
flask-cors
openai
python-dotenv
assemblyai